requirements.txt
db-dtypes>=1.1.1
google-cloud-aiplatform>=1.46.0
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0
ipykernel==6.29.3
matplotlib==3.8.4
numpy==1.22
pandas==1.3.5
plotly==5.18.0
streamlit==1.30.0
streamlit_feedback==0.1.3
seaborn==0.13.1
vertexai==1.43.0